Started implementing texture abstraction.

+++ engine/src/core/platform/opengl/textures/opengl_texture.cpp
@@ -40,7 +40,7 @@ namespace Bald::Platform::Graphics {
             glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
             glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
 
-            glTexImage2D(GL_TEXTURE_2D, 0, format, m_Width, m_Height, 0, format, GL_UNSIGNED_BYTE, m_Data);
+            glTexImage2D(GL_TEXTURE_2D, 0, static_cast<int32_t>(format), m_Width, m_Height, 0, format, GL_UNSIGNED_BYTE, m_Data);
             glGenerateMipmap(GL_TEXTURE_2D);
         } else {
             CORE_LOG_ERROR("[OpenGLTexture] Failed to load texture!");
